+++ fr.rostren.tracker.model/src-gen/fr/rostren/tracker/impl/OperationServiceImpl.java
@@ -38,7 +38,7 @@ public class OperationServiceImpl extends EObjectImpl implements OperationServic
 	 * @generated
 	 */
 	@Override
-	public float sumAmounts(Operation operation) {
+	public double sumAmounts(Operation operation) {
 		// TODO: implement this method
 		// Ensure that you remove @generated or mark it @generated NOT
 		throw new UnsupportedOperationException();
